refactor code to use cvector instead of linked lists

+++ src/token.h
@@ -1,18 +1,20 @@
 #ifndef BTC_TOKEN_H_
 #define BTC_TOKEN_H_
 
+#include <stddef.h>
+
 typedef struct _btc_token btc_token;
 typedef struct _btc_token_range btc_token_range;
 
 struct _btc_token_range {
 	// Start line number
-	int start_line_number;
+	size_t start_line_number;
 	// End line number
-	int end_line_number;
+	size_t end_line_number;
 	// Start offset
-	int start_offset;
+	size_t start_offset;
 	// End offset
-	int end_offset;
+	size_t end_offset;
 };
 
 struct _btc_token {
@@ -23,6 +25,7 @@ struct _btc_token {
     double number;
 
     struct _btc_token_range range;
+
 };
 
 void btc_destroy_token(btc_token* token);
